# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Music Platform Gateway API
  version: "2026-03-04"
  description: |
    Gateway API for the split architecture deployment.
    Base gateway: http://127.0.0.1:8080
    Verified with live tests on 2026-03-04.
servers:
  - url: http://127.0.0.1:8080
    description: Local gateway
  - url: http://slcdut.xyz:8080
    description: Public gateway
...
paths:
...
  /get_music:
    get:
      tags: [catalog]
      summary: Get music detail by path
      operationId: getMusicByPath
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
          example: hua_hai/hua_hai.mp3
      responses:
        "200":
          description: Music detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelope"
...
  /music/search:
    post:
      tags: [catalog]
      summary: Search music by keyword
      description: |
        Uses keyword-based routing while keeping the legacy response schema.
        If the keyword contains Chinese characters, local catalog search is tried first and Jamendo is used as fallback.
        If the keyword contains Latin letters and no Chinese characters, Jamendo is tried first and local catalog search is used as fallback.
        Jamendo results are mapped into `FileListItem` with a virtual `path` that can be sent back to `/stream` and `/download`.
      operationId: searchMusic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
            example:
              keyword: hai
      responses:
        "200":
          description: Music search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileListItem"
...
components:
  responses:
    ErrorEnvelope:
      description: Error response envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required: [code, message]
...
    SearchRequest:
      type: object
      required: [keyword]
      properties:
        keyword:
          type: string
...
    FileListItem:
      type: object
      description: Legacy search/list item. For Jamendo-compatible search results, `path` is a virtual filename rather than a local relative path.
      properties:
        path:
          type: string
        duration:
          type: string
        artist:
          type: string
        cover_art_url:
          type: string
          nullable: true
      required: [path, duration]
    MusicResponse:
      type: object
      description: Legacy playback response. `stream_url` may point to local uploads or Jamendo upstream audio depending on the requested filename/path.
      properties:
        stream_url:
          type: string
        lrc_url:
          type: string
          nullable: true
        album_cover_url:
          type: string
          nullable: true
        duration:
          type: number
          format: double
          nullable: true
        title:
          type: string
        artist:
          type: string
        album:
          type: string
      required: [stream_url]
...